Move all methods from LichSuDatVeDAO to DatVeDAO
